+++ src/crontab/tests/SchedulerTest.php
@@ -1,7 +1,16 @@
 <?php
 
-namespace HyperfTest\Crontab;
+declare(strict_types=1);
+/**
+ * This file is part of Hyperf.
+ *
+ * @link     https://www.hyperf.io
+ * @document https://doc.hyperf.io
+ * @contact  group@hyperf.io
+ * @license  https://github.com/hyperf-cloud/hyperf/blob/master/LICENSE
+ */
 
+namespace HyperfTest\Crontab;
 
 use Hyperf\Crontab\CrontabManager;
 use Hyperf\Crontab\Parser;
@@ -9,9 +18,12 @@ use Hyperf\Crontab\Scheduler;
 use PHPUnit\Framework\TestCase;
 use ReflectionMethod;
 
+/**
+ * @internal
+ * @coversNothing
+ */
 class SchedulerTest extends TestCase
 {
-
     public function testGetSchedules()
     {
         $scheduler = new Scheduler(new CrontabManager(new Parser()));
@@ -20,5 +32,4 @@ class SchedulerTest extends TestCase
         $result = $reflectionMethod->invoke($scheduler);
         $this->assertSame([], $result);
     }
-
 }