+++ src/task/src/Listener/OnFinishListener.php
@@ -16,6 +16,7 @@ use Hyperf\Event\Annotation\Listener;
 use Hyperf\Event\Contract\ListenerInterface;
 use Hyperf\Framework\Event\OnFinish;
 use Hyperf\Task\ChannelFactory;
+use Hyperf\Task\Finish;
 use Psr\Container\ContainerInterface;
 
 /**
@@ -42,9 +43,9 @@ class OnFinishListener implements ListenerInterface
 
     public function process(object $event)
     {
-        if ($event instanceof OnFinish) {
+        if ($event instanceof OnFinish && $event->data instanceof Finish) {
             $factory = $this->container->get(ChannelFactory::class);
-            $factory->push($event->taskId, $event->data);
+            $factory->push($event->taskId, $event->data->data);
         }
     }
 }